+++ easy-es-springboot-test/src/test/java/org/dromara/easyes/test/mix/MixTest.java
@@ -111,7 +111,7 @@ public class MixTest {
         wrapper.sort(SortOptions.of(a -> a.script(b -> b
                 .type(ScriptSortType.Number)
                 .order(SortOrder.Desc)
-                .script(c -> c.source("doc['star_num'].value"))
+                .script(c -> c.source(s -> s.scriptString("doc['star_num'].value")))
         )));
         List<Document> documents = documentMapper.selectList(wrapper);
         System.out.println(documents);